Module amazon_sp_product_pricing::apis::product_pricing_api

Enums

GetCompetitivePricingError

struct for typed errors of method get_competitive_pricing

GetItemOffersBatchError

struct for typed errors of method get_item_offers_batch

GetItemOffersError

struct for typed errors of method get_item_offers

GetListingOffersBatchError

struct for typed errors of method get_listing_offers_batch

GetListingOffersError

struct for typed errors of method get_listing_offers

GetPricingError

struct for typed errors of method get_pricing

Functions

get_competitive_pricing

Returns competitive pricing information for a seller’s offer listings based on seller SKU or ASIN. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 1 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_item_offers

Returns the lowest priced offers for a single item based on ASIN. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 1 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_item_offers_batch

Returns the lowest priced offers for a batch of items based on ASIN. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 1 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_listing_offers

Returns the lowest priced offers for a single SKU listing. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 2 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_listing_offers_batch

Returns the lowest priced offers for a batch of listings by SKU. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 1 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_pricing

Returns pricing information for a seller’s offer listings based on seller SKU or ASIN. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 1 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.